backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "媒体信息管理系统API后端"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pillow>=11.0.0",
    "opencv-python>=4.10.0",
]
